+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### Unreleased
 - Add healthcheck endpoint for service
 - Add asynchronous behaviour for client/server actions
-- Add loggign of message stats
+- Add logging of message stats
 
 ### 0.1.0 2017-03-29
 - Poll url healthcheck and aliveness endpoints