+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Unreleased
+
+### 1.0.0 2017-04-12
 - Add healthcheck endpoint for service
 - Add asynchronous behaviour for client/server actions
 - Add logging of message stats